+++ src/infra/instrumentation/instrumentation-agent.js
@@ -37,9 +37,10 @@ class InstrumentationAgent {
       this.raven = Raven.config(process.env.SENTRY_URL, {
         environment: process.env.HULL_ENV || "production",
         release: this.manifest.version,
-        captureUnhandledRejections: true
-      }).install((loggedInSentry, err = {}) => {
-        console.error("connector.error", { loggedInSentry, err: err.stack || err });
+        captureUnhandledRejections: false,
+        sampleRate: parseFloat(process.env.SENTRY_SAMPLE_RATE) || 1.0
+      }).install((err) => {
+        console.error("connector.error", { err: err.stack || err });
         if (this.exitOnError) {
           if (process.listenerCount("gracefulExit") > 0) {
             process.emit("gracefulExit");
@@ -47,6 +48,20 @@ class InstrumentationAgent {
             process.exit(1);
           }
         }
+      });
+
+      global.process.on("unhandledRejection", (reason, promise) => {
+        const context = promise.domain && promise.domain.sentryContext;
+        this.raven.captureException(reason, context || {}, () => {
+          console.error("connector.error", { reason });
+          if (this.exitOnError) {
+            if (process.listenerCount("gracefulExit") > 0) {
+              process.emit("gracefulExit");
+            } else {
+              process.exit(1);
+            }
+          }
+        });
       });
     }