+++ src/types/hull-segment.js
@@ -6,4 +6,7 @@
 export type THullSegment = {
   id: string;
   name: string;
+  stats: {
+    users: Number
+  };
 };