+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 0.13.11
+* this release brings bigger changes to error handling:
+  - it cleans up a little middleware stack including smart-notifier errors
+  - it introduces two types of errors - `unhandled error` which is handled the same as till now, and `transient error` which won't be pushed to sentry, but only instrumented in datadog
+  - it deprecates dedicated smartNotifierErrorMiddleware
+  - smartNotifierHandler in case of error behaves like notifHandler and pass the error down the middleware stack
+* added `timeout` option to `Hull.Connector` constructor to control the timeout value
+* upgrades `raven` library
+* flow types fixes
+
 # 0.13.10
 * from now we test each commit on multiple nodejs versions
 * in case of smart-notifier notification if requestId is not passed as an http header we fallback to notification_id from body